pyproject.toml
[tool.poetry]
name = "spanner-py-demo"
version = "0.1.0"
description = ""
authors = ["kazshinohara <kazshinohara@gcpx.org>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = "^0.17.4"
sqlalchemy-spanner = "^1.0.0"
SQLAlchemy-Utils = "^0.38.2"
alembic = "^1.7.6"
pytest = "^7.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
